Add satellite.py
satellite.py uses newton's method to determine the proximity of a simulated satellite to a desired location. It determines if the satellite is above the horizon and which four satellites in the array are best suited to calculate said location. It then sends the proper satellite data to the receiver via the vehicle.